solver can now extract addresses from namescript
