+++ src/primitives/transaction.h
@@ -16,6 +16,27 @@
 class CDiskTxPos;
 
 static const int NAMECOIN_TX_VERSION = 0x0666; //0x0666 is initial version
+typedef std::vector<unsigned char> CNameVal;
+struct NameTxInfo
+{
+    CNameVal name;
+    CNameVal value;
+    int nRentalDays;
+    int op;
+    int nOut;
+    std::string err_msg; //in case function that takes this as argument have something to say about it
+
+    //used only by DecodeNameScript()
+    std::string strAddress;
+    bool fIsMine;
+
+    //used only by GetNameList()
+    int nExpiresAt;
+
+    NameTxInfo(): nRentalDays(-1), op(-1), nOut(-1), fIsMine(false), nExpiresAt(-1) {}
+    NameTxInfo(CNameVal name, CNameVal value, int nRentalDays, int op, int nOut, std::string err_msg):
+        name(name), value(value), nRentalDays(nRentalDays), op(op), nOut(nOut), err_msg(err_msg), fIsMine(false), nExpiresAt(-1) {}
+};
 
 /** An outpoint - a combination of a transaction hash and an index n into its vout */
 class COutPoint
