+++ src/script/standard.h
@@ -63,6 +63,7 @@ enum txnouttype
     TX_SCRIPTHASH,
     TX_MULTISIG,
     TX_NULL_DATA,
+    TX_NAME
 };
 
 class CNoDestination {
