+++ src/wallet_ismine.h
@@ -23,8 +23,8 @@ enum isminetype
 /** used for bitflags of isminetype */
 typedef uint8_t isminefilter;
 
+isminetype IsMine(const CKeyStore& keystore, const CTxDestination& dest);
 isminetype IsMine(const CKeyStore& keystore, const CScript& scriptPubKey);
 isminetype IsMine(const CKeyStore& keystore, const CScript& scriptPubKey, bool& fName);
-isminetype IsMine(const CKeyStore& keystore, const CTxDestination& dest);
 
 #endif // BITCOIN_WALLET_ISMINE_H
